% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{cdk.version}
\alias{cdk.version}
\title{Get Version of the CDK Library}
\usage{
cdk.version()
}
\value{
A character string reporting the version of the CDK library used in this package
}
\description{
Get Version of the CDK Library
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
