% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\docType{class}
\name{cdkFormula-class}
\alias{cdkFormula-class}
\alias{show,cdkFormula-method}
\title{Class cdkFormula, ac class for handling molecular formula}
\description{
This class handles molecular formulae. It provides extra 
 information such as the IMolecularFormula Java object, elements contained 
 and number of them.
}
\section{Objects from the Class}{

  Objects can be created using new constructor and filled with a specific 
  mass and window accuracy
}

\references{
A parallel effort to expand the Chemistry Development 
  Kit: \url{https://cdk.github.io/}
}
\seealso{
\code{\link{get.formula}}
\code{\link{set.charge.formula}}
\code{\link{get.isotopes.pattern}}
\code{\link{isvalid.formula}}
}
\author{
Miguel Rojas-Cherto (\email{miguelrojasch@yahoo.es})
}
