\name{cdk.version}
\alias{cdk.version}
\title{
  Get Current CDK Version
}
\description{
Returns a string containing the version of the CDK used in this package
}
\usage{
cdk.version()
}
\value{
A string representing the CDK version
}
\keyword{programming}
\author{Rajarshi Guha (\email{rajarshi.guha@gmail.com})}
