% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{compare.isotope.pattern}
\alias{compare.isotope.pattern}
\title{Compare isotope patterns.}
\usage{
compare.isotope.pattern(iso1, iso2, ips = NULL)
}
\arguments{
\item{iso1}{The first isotope pattern, which should be a \code{jobjRef} corresponding to the \code{IsotopePattern} class}

\item{iso2}{The second isotope pattern, which should be a \code{jobjRef} corresponding to the \code{IsotopePattern} class}

\item{ips}{An instance of the \code{IsotopePatternSimilarity} class. if \code{NULL} one will be constructed automatically}
}
\value{
A numeric value between 0 and 1 indicating the similarity between the two patterns
}
\description{
Computes a similarity score between two different isotope abundance patterns.
}
\references{
\url{https://cdk.github.io/cdk/2.10/docs/api/org/openscience/cdk/formula/IsotopePatternSimilarity.html}
}
\seealso{
\code{\link{get.isotope.pattern.similarity}}
}
\author{
Miguel Rojas Cherto
}
