% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{copy.image.to.clipboard}
\alias{copy.image.to.clipboard}
\title{copy.image.to.clipboard}
\usage{
copy.image.to.clipboard(molecule, depictor = NULL)
}
\arguments{
\item{molecule}{The molecule to query. Should be a `jobjRef` representing an `IAtomContainer`}

\item{depictor}{Optional. Default \code{NULL}. Depictor from \code{get.depictor}}
}
\description{
generate an image and make it available to the system
clipboard.
}
