% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{do.aromaticity}
\alias{do.aromaticity}
\title{do.aromaticity}
\usage{
do.aromaticity(mol)
}
\arguments{
\item{mol}{The molecule to query. Should be a `jobjRef` representing an `IAtomContainer`}
}
\description{
detect aromoticity of an input compound
}
