% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{do.isotopes}
\alias{do.isotopes}
\title{do.isotopes}
\usage{
do.isotopes(mol)
}
\arguments{
\item{mol}{The molecule to query. Should be a `jobjRef` representing an `IAtomContainer`}
}
\description{
configure isotopes
}
