% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{do.typing}
\alias{do.typing}
\title{do.typing}
\usage{
\method{do}{typing}(mol)
}
\arguments{
\item{mol}{The molecule to query. Should be a `jobjRef` representing an `IAtomContainer`}
}
\description{
configure atom typings.
}
