% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{get.atom.count}
\alias{get.atom.count}
\title{Get the number of atoms in the molecule.}
\usage{
get.atom.count(mol)
}
\arguments{
\item{mol}{The molecule to query. Should be a `jobjRef` representing an `IAtomContainer`}
}
\value{
An integer representing the number of atoms in the molecule
}
\description{
Get the number of atoms in the molecule.
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
