% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{get.atom.index}
\alias{get.atom.index}
\title{get.atom.index}
\usage{
get.atom.index(atom, mol)
}
\arguments{
\item{atom}{The atom object}

\item{mol}{The `IAtomContainer` object containing the atom}
}
\value{
An integer representing the atom index.
}
\description{
Get the index of an atom in a molecule.
}
\details{
Access the index of an atom in the context of an IAtomContainer. 
Indexing starts from 0. If the index is not known, -1 is returned.
}
\seealso{
\code{\link{get.connected.atom}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
