% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{get.atomic.desc.names}
\alias{get.atomic.desc.names}
\title{Get class names for atomic descriptors}
\usage{
get.atomic.desc.names()
}
\value{
A character vector containing class names for atomic descriptors
}
\description{
Get class names for atomic descriptors
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
