% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{get.atomic.number}
\alias{get.atomic.number}
\title{get.atomic.number}
\usage{
get.atomic.number(atom)
}
\arguments{
\item{atom}{The atom to query}
}
\value{
An integer representing the atomic number
}
\description{
Get the atomic number of the atom.
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
