% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{get.atoms}
\alias{get.atoms}
\title{Get the atoms from a molecule or bond.}
\usage{
get.atoms(object)
}
\arguments{
\item{object}{A `jobjRef` representing either a molecule (`IAtomContainer`) or
bond (`IBond`) object.}
}
\value{
A list of `jobjRef` representing the `IAtom` objects in the molecule or bond
}
\description{
Get the atoms from a molecule or bond.
}
\seealso{
\code{\link{get.bonds}}, \code{\link{get.connected.atoms}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
