% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{get.charge}
\alias{get.charge}
\title{Get the charge on the atom.

This method returns the partial charge on the atom. If charges have not been set the
return value is \code{NULL}, otherwise the appropriate charge.}
\usage{
get.charge(atom)
}
\arguments{
\item{atom}{The atom to query}
}
\value{
An numeric representing the partial charge. If charges have not been set,
`NULL` is returned
}
\description{
Get the charge on the atom.

This method returns the partial charge on the atom. If charges have not been set the
return value is \code{NULL}, otherwise the appropriate charge.
}
\seealso{
\code{\link{get.formal.charge}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
