% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{get.chem.object.builder}
\alias{get.chem.object.builder}
\title{Get the default chemical object builder.}
\usage{
get.chem.object.builder()
}
\value{
An instance of \href{http://cdk.github.io/cdk/2.2/docs/api/org/openscience/cdk/silent/SilentChemObjectBuilder.html}{SilentChemObjectBuilder}
}
\description{
The CDK employs a builder design pattern to construct
instances of new chemical objects (e.g., atoms, bonds, parsers
and so on). Many methods require an instance of a builder 
object to function. While most functions in this package handle
this internally, it is useful to be able to get an instance of
a builder object when directly working with the CDK API via
`rJava`.
}
\details{
This method returns an instance of the \href{http://cdk.github.io/cdk/2.2/docs/api/org/openscience/cdk/silent/SilentChemObjectBuilder.html}{SilentChemObjectBuilder}. 
Note that this is a static object that is created at package load time, 
and the same instance is returned whenever this function is called.
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
