% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{get.desc.categories}
\alias{get.desc.categories}
\title{List available descriptor categories}
\usage{
get.desc.categories()
}
\value{
A character vector listing available descriptor categories. This can be
used in \link{get.desc.names}
}
\description{
List available descriptor categories
}
\seealso{
\link{get.desc.names}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
