% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{get.exact.mass}
\alias{get.exact.mass}
\title{get.exact.mass}
\usage{
get.exact.mass(mol)
}
\arguments{
\item{mol}{The molecule to query. Should be a `jobjRef` representing an `IAtomContainer`}
}
\description{
get.exact.mass
}
