% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{get.formal.charge}
\alias{get.formal.charge}
\title{Get the formal charge on the atom.

By default the formal charge will be 0 (i.e., \code{NULL} is never returned).}
\usage{
get.formal.charge(atom)
}
\arguments{
\item{atom}{The atom to query}
}
\value{
An integer representing the formal charge
}
\description{
Get the formal charge on the atom.

By default the formal charge will be 0 (i.e., \code{NULL} is never returned).
}
\seealso{
\code{\link{get.charge}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
