% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{get.hydrogen.count}
\alias{get.hydrogen.count}
\title{Get the implicit hydrogen count for the atom.}
\usage{
get.hydrogen.count(atom)
}
\arguments{
\item{atom}{The atom to query}
}
\value{
An integer representing the hydrogen count
}
\description{
This method returns the number of implicit H's on the atom. 
Depending on where the molecule was read from this may be \code{NULL} or an integer
greater than or equal to 0
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
