% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{get.isotope.pattern.generator}
\alias{get.isotope.pattern.generator}
\title{Construct an isotope pattern generator.}
\usage{
get.isotope.pattern.generator(minAbundance = NULL)
}
\arguments{
\item{minAbundance}{The minimum abundance}
}
\value{
A \code{jobjRef} corresponding to an instance of \code{IsotopePatternGenerator}
}
\description{
Constructs an instance of the CDK \code{IsotopePatternGenerator}, with an optional
minimum abundance specified. This object can be used to generate all combinatorial
chemical isotopes given a structure.
}
\references{
\url{http://cdk.github.io/cdk/1.5/docs/api/org/openscience/cdk/formula/IsotopePatternGenerator.html}
}
\author{
Miguel Rojas Cherto
}
