% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{get.isotope.pattern.similarity}
\alias{get.isotope.pattern.similarity}
\title{Construct an isotope pattern similarity calculator.}
\usage{
get.isotope.pattern.similarity(tol = NULL)
}
\arguments{
\item{tol}{The tolerance}
}
\value{
A \code{jobjRef} corresponding to an instance of \code{IsotopePatternSimilarity}
}
\description{
A method that returns an instance of the CDK \code{IsotopePatternSimilarity}
class which can be used to compute similarity scores between pairs of
isotope abundance patterns.
}
\author{
Miguel Rojas Cherto
}
\references{
\url{http://cdk.github.io/cdk/1.5/docs/api/org/openscience/cdk/formula/IsotopePatternSimilarity.html}
}
\seealso{
\code{\link{compare.isotope.pattern}}
}

