% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{get.isotopes.pattern}
\alias{get.isotopes.pattern}
\title{get.isotopes.pattern}
\usage{
get.isotopes.pattern(formula, minAbund = 0.1)
}
\arguments{
\item{formula}{Required. A CDK molecule formula}

\item{minAbund}{Optional. Default \code{0.1}}
}
\description{
Generate the isotope pattern given a formula class
}
