% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{get.mol2formula}
\alias{get.mol2formula}
\title{get.mol2formula}
\usage{
get.mol2formula(molecule, charge = 0)
}
\arguments{
\item{molecule}{The molecule to query. Should be a `jobjRef` representing an `IAtomContainer`}

\item{charge}{Optional. Default \code{0}}
}
\description{
get.mol2formula
}
