% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{get.point2d}
\alias{get.point2d}
\title{Get the 2D coordinates of the atom.}
\usage{
get.point2d(atom)
}
\arguments{
\item{atom}{The atom to query}
}
\value{
A 2-element numeric vector representing the X & Y coordinates.
}
\description{
In case, coordinates are unavailable (e.g., molecule was read in from a 
SMILES file) or have not been generated yet, `NA`'s are returned for the 
X & Y coordinates.
}
\seealso{
\code{\link{get.point3d}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
