% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{get.point3d}
\alias{get.point3d}
\title{Get the 3D coordinates of the atom.

In case, coordinates are unavailable (e.g., molecule was read in from a 
SMILES file) or have not been generated yet, `NA`'s are returned for the 
X, Y and Z coordinates.}
\usage{
get.point3d(atom)
}
\arguments{
\item{atom}{The atom to query}
}
\value{
A 3-element numeric vector representing the X, Y and Z coordinates.
}
\description{
Get the 3D coordinates of the atom.

In case, coordinates are unavailable (e.g., molecule was read in from a 
SMILES file) or have not been generated yet, `NA`'s are returned for the 
X, Y and Z coordinates.
}
\examples{
\dontrun{
atoms <- get.atoms(mol)
coords <- do.call('rbind', lapply(apply, get.point3d))
}
}
\seealso{
\code{\link{get.point2d}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
