% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smiles.R
\name{get.smiles.parser}
\alias{get.smiles.parser}
\title{Get a SMILES parser object.}
\usage{
get.smiles.parser()
}
\value{
A `jobjRef` object corresponding to the CDK 
\href{http://cdk.github.io/cdk/2.2/docs/api/org/openscience/cdk/smiles/SmilesParser.html}{SmilesParser} class
}
\description{
This function returns a reference to a SMILES parser
object. If you are parsing multiple SMILES strings using multiple
calls to \code{\link{parse.smiles}}, it is
preferable to create your own parser and supply it to
\code{\link{parse.smiles}} rather than forcing that function
to instantiate a new parser for each call
}
\seealso{
\code{\link{get.smiles}}, \code{\link{parse.smiles}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
