% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereo.R
\name{get.stereocenters}
\alias{get.stereocenters}
\title{Identify which atoms are stereocenters.}
\usage{
get.stereocenters(mol)
}
\arguments{
\item{mol}{A \code{jObjRef} representing an \code{IAtomContainer}}
}
\value{
A logical vector of length equal in length to the number of atoms. The i'th element is \code{TRUE} if the i'th element is identified as a stereocenter
}
\description{
This method identifies stereocenters based on connectivity.
}
\seealso{
\code{\link{get.element.types}}, \code{\link{get.stereo.types}}
}
\author{
Rajarshi Guha \email{rajarshi.guha@gmail.com}
}
