% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{get.symbol}
\alias{get.symbol}
\title{Get the atomic symbol of the atom.

Get the atomic symbol of the atom.}
\usage{
get.symbol(atom)
}
\arguments{
\item{atom}{The atom to query}
}
\value{
A character representing the atomic symbol
}
\description{
Get the atomic symbol of the atom.

Get the atomic symbol of the atom.
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
