% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{get.title}
\alias{get.title}
\title{Get the title of the molecule.}
\usage{
get.title(mol)
}
\arguments{
\item{mol}{The molecule to query. Should be a `jobjRef` representing an `IAtomContainer`}
}
\value{
A character string with the title, `NA` is no title is specified
}
\description{
Some molecules may not have a title (such as when parsing in a SMILES
with not title).
}
\seealso{
\code{\link{set.title}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
