% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charges.R
\name{get.total.charge}
\alias{get.total.charge}
\title{Get the Total Charge for the Molecule}
\usage{
get.total.charge(mol)
}
\arguments{
\item{mol}{A `jobjRef` objects of Java class `IAtomContainer`}
}
\value{
An integer value indicating the total charge
}
\description{
returns the summed partial charges for a molecule
and \code{\link{get.total.formal.charge}} returns the summed formal charges. Currently,
if one or more partial charges are unset, the function simply returns
the sum of formal charges (via \code{\link{get.total.formal.charge}}). This is slightly
different from how the CDK evaluates the total charge of a molecule 
(via \href{https://cdk.github.io/cdk/2.2/docs/api/org/openscience/cdk/tools/manipulator/AtomContainerManipulator.html}{AtomContainerManipulator.getTotalCharge()}), 
but is in line with how OEChem determines net charge on a molecule. In general, you will 
want to use the \code{\link{get.total.charge}} function.
}
\section{Functions}{
\itemize{
\item \code{get.total.charge}: Get sum of partial charges
}}

\section{XXX}{
NA
}

\seealso{
\code{\link{get.total.formal.charge}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
