% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{get.total.hydrogen.count}
\alias{get.total.hydrogen.count}
\title{Get the Total Hydrogen Count for a Molecule}
\usage{
get.total.hydrogen.count(mol)
}
\arguments{
\item{mol}{A `jobjRef` objects of Java class `IAtomContainer`}
}
\value{
An integer value indicating the number of implicit hydrogens
}
\description{
The function will return the summed implicit hydrogens of all atoms in
the specified  `IAtomContainer`
}
\seealso{
\code{\link{remove.hydrogens}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
