% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{get.tpsa}
\alias{get.tpsa}
\title{Compute TPSA for a molecule}
\usage{
get.tpsa(molecule)
}
\arguments{
\item{molecule}{A molecule object}
}
\value{
A double value representing the TPSA value
}
\description{
Compute TPSA for a molecule
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
