% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{get.volume}
\alias{get.volume}
\title{Compute volume of a molecule}
\usage{
get.volume(molecule)
}
\arguments{
\item{molecule}{A molecule object}
}
\value{
A double value representing the volume
}
\description{
This method does not require 3D coordinates. As a result its an
approximation
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
