% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{get.xlogp}
\alias{get.xlogp}
\title{Compute XLogP for a molecule}
\usage{
get.xlogp(molecule)
}
\arguments{
\item{molecule}{A molecule object}
}
\value{
A double value representing the XLogP value
}
\description{
Compute XLogP for a molecule
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
