\name{generate.2d.coordinates}
\alias{generate.2d.coordinates}
\title{
 Generate 2D Coordinates from Connectivity Information	
}
\description{
 This function will generate reasonable 2D coordinates based purely on
 connectivity information.
}
\usage{
 generate.2d.coordinates(molecule)
}
\arguments{
  \item{molecule}{An \code{IAtomContainer} object that can be obtained
    by loading them from disk or drawing them in the editor.
  }
}
\value{
 Returns the input molecule with 2D coordinates added.	
}
\keyword{programming}
\author{Rajarshi Guha (\email{rajarshi.guha@gmail.com})}
