\name{get.total.hydrogen.count}
\alias{get.total.hydrogen.count}
\title{
  Get the Total Hydrogen Count for a Molecule
}
\description{
  The function will return the summed implicit hydrogens of all atoms in
  the specified  AtomContainer
}
\usage{
get.total.hydrogen.count(molecule)
}
\arguments{
  \item{molecule}{A Java object of class \code{IAtomContainer}}
}
\value{
  An integer value indicating the number of implicit hydrogens
}
\keyword{programming}

\author{Rajarshi Guha (\email{rajarshi.guha@gmail.com})}
