\name{get.total.charge}
\alias{get.total.charge}
\title{
  Get the Total Charge for the Molecule
}
\description{
  The function will return the summed charges of all atoms in
  the specified  AtomContainer
}
\usage{
get.total.charge(molecule)
}
\arguments{
  \item{molecule}{A Java object of class \code{IAtomContainer}}
}
\value{
  An double value indicating the total charge
}
\keyword{programming}

\author{Rajarshi Guha (\email{rajarshi.guha@gmail.com})}
