% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{is.aliphatic}
\alias{is.aliphatic}
\title{Tests whether an atom is aliphatic.

This assumes that the molecule containing the atom has been 
appropriately configured.}
\usage{
is.aliphatic(atom)
}
\arguments{
\item{atom}{The atom to test}
}
\value{
`TRUE` is the atom is aliphatic, `FALSE` otherwise
}
\description{
Tests whether an atom is aliphatic.

This assumes that the molecule containing the atom has been 
appropriately configured.
}
\seealso{
\code{\link{is.in.ring}}, \code{\link{is.aromatic}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
