% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{is.aromatic}
\alias{is.aromatic}
\title{is.aromatic}
\usage{
is.aromatic(atom)
}
\arguments{
\item{atom}{The atom to test}
}
\value{
`TRUE` is the atom is aromatic, `FALSE` otherwise
}
\description{
Tests whether an atom is aromatic.
}
\details{
This assumes that the molecule containing the atom has been 
appropriately configured.
}
\seealso{
\code{\link{is.aliphatic}}, \code{\link{is.in.ring}}, \code{\link{do.aromaticity}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
