% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{is.in.ring}
\alias{is.in.ring}
\title{Tests whether an atom is in a ring.}
\usage{
is.in.ring(atom)
}
\arguments{
\item{atom}{The atom to test}
}
\value{
`TRUE` is the atom is in a ring, `FALSE` otherwise
}
\description{
This assumes that the molecule containing the atom has been 
appropriately configured.
}
\seealso{
\code{\link{is.aliphatic}}, \code{\link{is.aromatic}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
