% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{is.neutral}
\alias{is.neutral}
\alias{charge}
\title{Tests whether the molecule is neutral.}
\usage{
is.neutral(mol)
}
\arguments{
\item{mol}{The molecule to query. Should be a `jobjRef` representing an `IAtomContainer`}
}
\value{
`TRUE` if molecule is neutral, `FALSE` otherwise
}
\description{
The test checks whether all atoms in the molecule have a formal charge of 0.
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
