% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{matches}
\alias{matches}
\title{matches}
\usage{
matches(query, target, return.matches = FALSE)
}
\arguments{
\item{query}{Required. A SMARTSQuery}

\item{target}{Required. The molecule to query. Should be a `jobjRef` representing an `IAtomContainer`}

\item{return.matches}{Optional. Default \code{FALSE}}
}
\description{
matches
}
