% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\docType{package}
\name{rcdk-deprecated}
\alias{-package}
\alias{rcdk-deprecated}
\alias{deprecated_rcdk_function}
\alias{do.typing}
\title{Deprecated functions in the rcdk package.}
\usage{
deprecated_rcdk_function(x, value, ...)
}
\arguments{
\item{x}{For assignment operators, the object that will undergo a replacement
(object inside parenthesis).}

\item{value}{For assignment operators, the value to replace with 
(the right side of the assignment).}

\item{...}{For functions other than assignment operators, 
parameters to be passed to the modern version of the function (see table).}
}
\description{
These functions are provided for compatibility with older version of
the phyloseq package.  They may eventually be completely
removed.
}
\details{
\tabular{rl}{
  \code{do.typing} \tab now a synonym for \code{\link{set.atom.types}}\cr
}
}
