% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{remove.hydrogens}
\alias{remove.hydrogens}
\title{Remove Hydrogens from a Molecule}
\usage{
remove.hydrogens(mol)
}
\arguments{
\item{mol}{A `jobjRef` object of Java class `IAtomContainer`}
}
\value{
A `jobjRef` object of Java class `IAtomContainer`, with no hydrogens
}
\description{
This function generate a new `IAtomContainer`` object in which the
hydrogens have been removed. This can be useful for descriptor
calculations.
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
