% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{set.atom.types}
\alias{set.atom.types}
\title{set.atom.types}
\usage{
set.atom.types(mol)
}
\arguments{
\item{mol}{The molecule whose atoms should be typed}
}
\value{
Nothing is returned, the molecule is modified in place
}
\description{
Set the CDK atom types for all atoms in the molecule.
}
\details{
Calling this method will overwrite any pre-existing type information. 
Currently there is no way to choose other atom typing schemes
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
