% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{set.title}
\alias{set.title}
\title{Set the title of the molecule.}
\usage{
set.title(mol, title = "")
}
\arguments{
\item{mol}{The molecule to query. Should be a `jobjRef` representing an `IAtomContainer`}

\item{title}{The title of the molecule as a character string. This will overwrite
any pre-existing title. The default value is an empty string.}
}
\description{
Set the title of the molecule.
}
\seealso{
\code{\link{get.title}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
