% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{view.image.2d}
\alias{view.image.2d}
\title{view.image.2d}
\usage{
view.image.2d(molecule, depictor = NULL)
}
\arguments{
\item{molecule}{The molecule to display Should be a `jobjRef` representing an `IAtomContainer`}

\item{depictor}{Default \code{NULL}}
}
\description{
view.image.2d
}
