% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{view.molecule.2d}
\alias{view.molecule.2d}
\title{view.molecule.2d}
\usage{
view.molecule.2d(
  molecule,
  ncol = 4,
  width = 200,
  height = 200,
  depictor = NULL
)
}
\arguments{
\item{molecule}{The molecule to query. Should be a `jobjRef` representing an `IAtomContainer`}

\item{ncol}{Default \code{4}}

\item{width}{Default \code{200}}

\item{height}{Default \code{200}}

\item{depictor}{Default \code{NULL}}
}
\description{
Create a 2D depiction of a molecule. If there are more than
one molecules supplied, return a grid with \code{ncol} columns,.
}
