% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{view.table}
\alias{view.table}
\title{view.table}
\usage{
view.table(molecules, dat, depictor = NULL)
}
\arguments{
\item{molecules}{A list of molecule objects (`jobjRef` representing an `IAtomContainer`)}

\item{dat}{The \code{data.frame} associated with the molecules, one per row}

\item{depictor}{Default \code{NULL}}
}
\description{
Create a tabular view of a set of molecules (in 2D) and associated data columns
}
