% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bpdata}
\alias{bpdata}
\title{Boiling Point Data}
\format{A data frame with 277 rows and 2 columns.:
\describe{
  \item{SMILES}{Structure in SMILES format}
  \item{BP}{Boiling point in Kelvin}
  
} The names of the molecules are used as the row names.}
\usage{
bpdata
}
\description{
A dataset containing the structures and associated boiling 
points for 277 molecules, primarily alkanes and substituted alkanes.
}
\references{
Goll, E.S. and Jurs, P.C.; "Prediction of the Normal Boiling Points of
 Organic Compounds From Molecular Structures with a Computational
 Neural Network Model", \emph{J. Chem. Inf. Comput. Sci.}, 1999, 
 \emph{39}, 974-983.
}
\keyword{datasets}
