% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{cdk.version}
\alias{cdk.version}
\title{Get the current CDK version used in the package.}
\usage{
cdk.version()
}
\value{
Returns a character containing the version of the CDK used in this package
}
\description{
Get the current CDK version used in the package.
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
