\name{cdkFormula-class}
\docType{class}
\alias{cdkFormula-class}
\alias{show,cdkFormula-method}

\title{Class cdkFormula, a class for handling molecular formula}
\description{  
  This class handles molecular formulae. It provides extra information
  such as the \code{IMolecularFormula} Java object, elements contained and
  number of them.
}
\section{Objects from the Class}{
  Objects can be created using \code{new}
  constructor and filled with a specific mass and window accuracy.  
}
\note{
  No notes yet.
}

\seealso{
  \code{\link{get.formula}},
  \code{\link{set.charge.formula}},
  \code{\link{get.isotopes.pattern}},
  \code{\link{isvalid.formula}},
}
\keyword{classes}
\author{Miguel Rojas-Cherto (\email{miguelrojasch@yahoo.es})}
\references{
  A parallel effort to expand the Chemistry Development Kit:
  \url{http://cdk.sourceforge.net}
}
