% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{generate.formula}
\alias{generate.formula}
\title{generate.formula}
\usage{
generate.formula(
  mass,
  window = 0.01,
  elements = list(c("C", 0, 50), c("H", 0, 50), c("N", 0, 50), c("O", 0, 50), c("S", 0,
    50)),
  validation = FALSE,
  charge = 0
)
}
\description{
generate.formula
}
