% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{get.bonds}
\alias{get.bonds}
\title{Get the bonds in a molecule.}
\usage{
get.bonds(mol)
}
\arguments{
\item{mol}{A `jobjRef` representing the molecule (`IAtomContainer`) object.}
}
\value{
A list of `jobjRef` representing the bonds (`IBond`) objects in the molecule
}
\description{
Get the bonds in a molecule.
}
\seealso{
\code{\link{get.atoms}}, \code{\link{get.connected.atoms}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
