% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{get.connected.atoms}
\alias{get.connected.atoms}
\title{Get atoms connected to the specified atom

Returns a `list`` of atoms that are connected to the specified atom.}
\usage{
get.connected.atoms(atom, mol)
}
\arguments{
\item{atom}{The atom object}

\item{mol}{The `IAtomContainer` object containing the atom}
}
\value{
A `list` containing `IAtom` objects, representing the atoms directly
connected to the specified atom
}
\description{
Get atoms connected to the specified atom

Returns a `list`` of atoms that are connected to the specified atom.
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
