% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{get.depictor}
\alias{get.depictor}
\title{get.depictor}
\usage{
get.depictor(
  width = 200,
  height = 200,
  zoom = 1.3,
  style = "cow",
  annotate = "off",
  abbr = "on",
  suppressh = TRUE,
  showTitle = FALSE,
  smaLimit = 100,
  sma = NULL,
  fillToFit = FALSE
)
}
\arguments{
\item{width}{Default. \code{200}}

\item{height}{Default. \code{200}}

\item{zoom}{Default. \code{1.3}}

\item{style}{Default. \code{cow}}

\item{annotate}{Default. \code{off}}

\item{abbr}{Default. \code{on}}

\item{suppressh}{Default. \code{TRUE}}

\item{showTitle}{Default. \code{FALSE}}

\item{smaLimit}{Default. \code{100}}

\item{sma}{Default. \code{NULL}}

\item{fillToFit}{Defailt. \code{FALSE}}
}
\description{
return an RcdkDepictor.
}
