% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{get.total.charge}
\alias{get.total.charge}
\title{get.total.charge}
\usage{
get.total.charge(mol)
}
\arguments{
\item{mol}{The molecule to query. Should be a `jobjRef` representing an `IAtomContainer`}
}
\description{
get.total.charge
}
