% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{get.total.hydrogen.count}
\alias{get.total.hydrogen.count}
\title{Get total number of implicit hydrogens in the molecule.}
\usage{
get.total.hydrogen.count(mol)
}
\arguments{
\item{mol}{The molecule to query. Should be a `jobjRef` representing an `IAtomContainer`}
}
\value{
An integer representing the total number of implicit hydrogens
}
\description{
Counts the number of hydrogens on the provided molecule. As this method
will sum all implicit hydrogens on each atom it is important to ensure
the molecule has already been configured (and thus each atom has an
implicit hydrogen count).
}
\seealso{
\code{\link{get.hydrogen.count}}, \code{\link{remove.hydrogens}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
