% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdk.R
\name{remove.hydrogens}
\alias{remove.hydrogens}
\title{Remove explicit hydrogens.}
\usage{
remove.hydrogens(mol)
}
\arguments{
\item{mol}{The molecule to query. Should be a `jobjRef` representing an `IAtomContainer`}
}
\value{
A copy of the original molecule, with explicit hydrogens removed
}
\description{
Create an copy of the original structure with explicit hydrogens removed. 
Stereochemistry is updated but up and down bonds in a depiction may need 
to be recalculated. This can also be useful for descriptor calculations.
}
\seealso{
\code{\link{get.hydrogen.count}}, \code{\link{get.total.hydrogen.count}}
}
\author{
Rajarshi Guha (\email{rajarshi.guha@gmail.com})
}
